// frontend/src/shared/lib/api/model/importStatus.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * prot-cellar
 * OpenAPI spec version: 0.1.0
 */

export type ImportStatus = typeof ImportStatus[keyof typeof ImportStatus];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const ImportStatus = {
  queued: 'queued',
  running: 'running',
  succeeded: 'succeeded',
  failed: 'failed',
  cancelled: 'cancelled',
} as const;