// frontend/src/shared/lib/api/model/startImportBody.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * prot-cellar
 * OpenAPI spec version: 0.1.0
 */
import type { ImportType } from './importType';
import type { StartImportBodyParams } from './startImportBodyParams';

export interface StartImportBody {
  import_type: ImportType;
  params?: StartImportBodyParams;
}